+++ flask/ctx.py
@@ -122,7 +122,8 @@ def copy_current_request_context(f):
     """A helper function that decorates a function to retain the current
     request context.  This is useful when working with greenlets.  The moment
     the function is decorated a copy of the request context is created and
-    then pushed when the function is called.
+    then pushed when the function is called.  A copy of the current session is
+    also included in the copied request context.
 
     Example::
 
@@ -133,13 +134,17 @@ def copy_current_request_context(f):
         def index():
             @copy_current_request_context
             def do_some_work():
-                # do some work here, it can access flask.request like you
-                # would otherwise in the view function.
+                # do some work here, it can access flask.request or
+                # flask.session like you would otherwise in the view function.
                 ...
             gevent.spawn(do_some_work)
             return 'Regular response'
 
     .. versionadded:: 0.10
+
+    .. versionchanged:: 1.0.3
+       A copy of the current session object is added to the request context
+       copy. This prevents `flask.session` pointing to an out-of-date object.
     """
     top = _request_ctx_stack.top
     if top is None: